# backend/swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TEAM WORKS API",
    "description": "팀 단위 캘린더 기반 일정관리 + 채팅 애플리케이션 API. API Route 소스 위치: backend/app/api/. v1.7.0 — Postits/Notices/Projects/ProjectSchedules/SubSchedules/Me/Refresh paths 보강",
    "version": "1.7.1",
    "x-source-directory": "backend/app/api"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Base URL (Next.js API Routes — backend/app/api/)"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "에러 메시지"
          }
        }
      },
...
      "Message": {
        "type": "object",
        "required": [
          "id",
          "teamId",
          "type",
          "senderId",
          "senderName",
          "content",
          "sentAt",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "a7b8c9d0-e1f2-3456-abcd-ef7890123456"
          },
          "teamId": {
            "type": "string",
            "format": "uuid",
            "example": "b1c2d3e4-f5a6-7890-bcde-fa1234567890"
          },
          "projectId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "프로젝트 채팅일 경우 소속 프로젝트 UUID, 팀 일자별이면 null",
            "example": null
          },
          "type": {
            "type": "string",
            "enum": [
              "NORMAL",
              "WORK_PERFORMANCE"
            ],
            "example": "NORMAL"
          },
          "senderId": {
            "type": "string",
            "format": "uuid",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "senderName": {
            "type": "string",
            "example": "홍길동"
          },
          "content": {
            "type": "string",
            "example": "오늘 미팅 시간 변경합니다. 오후 3시로 조정해주세요."
          },
          "sentAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-04-07T01:30:00.000Z"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-04-07T01:30:00.000Z"
          }
        }
      },
...
      "BoardAttachment": {
        "type": "object",
        "required": [
          "id",
          "postId",
          "originalName",
          "mimeType",
          "sizeBytes",
          "uploadedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "c4d5e6f7-a8b9-0123-cdef-ab4567890123"
          },
          "postId": {
            "type": "string",
            "format": "uuid",
            "example": "b3c4d5e6-f7a8-9012-bcde-fa3456789012"
          },
          "originalName": {
            "type": "string",
            "description": "업로드 시 원본 파일명",
            "example": "project-plan.pdf"
          },
          "mimeType": {
            "type": "string",
            "description": "검증된 MIME 타입",
            "example": "application/pdf"
          },
          "sizeBytes": {
            "type": "integer",
            "description": "파일 크기 (바이트)",
            "example": 204800
          },
          "uploadedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-04-20T01:00:00.000Z"
          }
        }
      },
      "BoardPost": {
        "type": "object",
        "required": [
          "id",
          "teamId",
          "authorId",
          "authorName",
          "title",
          "content",
          "createdAt",
          "updatedAt",
          "attachments"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "b3c4d5e6-f7a8-9012-bcde-fa3456789012"
          },
          "teamId": {
            "type": "string",
            "format": "uuid",
            "example": "b1c2d3e4-f5a6-7890-bcde-fa1234567890"
          },
          "projectId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "프로젝트 자료실이면 해당 UUID, 팀 일자별이면 null",
            "example": null
          },
          "authorId": {
            "type": "string",
            "format": "uuid",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "authorName": {
            "type": "string",
            "description": "작성자 표시 이름",
            "example": "홍길동"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "example": "프로젝트 계획서"
          },
          "content": {
            "type": "string",
            "maxLength": 20000,
            "example": "프로젝트 초기 계획서를 첨부합니다."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-04-20T01:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-04-20T01:00:00.000Z"
          },
          "attachments": {
            "type": "array",
            "description": "첨부파일 메타데이터 목록 (없으면 빈 배열)",
            "items": {
              "$ref": "#/components/schemas/BoardAttachment"
            }
          }
        }
      },
      "CreatePostRequest": {
        "type": "object",
        "description": "자료실 글 생성 요청 (multipart/form-data)",
        "required": [
          "title"
        ],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "description": "게시글 제목",
            "example": "프로젝트 계획서"
          },
          "content": {
            "type": "string",
            "maxLength": 20000,
            "description": "게시글 본문",
            "example": "프로젝트 초기 계획서를 첨부합니다."
          },
          "projectId": {
            "type": "string",
            "format": "uuid",
            "description": "프로젝트 자료실로 저장 시 지정 (선택)",
            "example": "e1f2a3b4-c5d6-7890-efab-cd1234567890"
          },
          "file": {
            "type": "string",
            "format": "binary",
            "description": "첨부파일 (단일, 최대 10MB, BR-09 MIME 화이트리스트)"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/teams/{teamId}/projects/{projectId}/messages": {
      "get": {
        "tags": [
          "ProjectMessages"
        ],
        "summary": "프로젝트 채팅 메시지 조회",
        "description": "프로젝트 전용 채팅 메시지를 조회합니다. sentAt 오름차순, 최대 200건. WORK_PERFORMANCE 타입은 LEADER 또는 업무보고 권한 보유자만 조회 가능합니다.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "소속 팀 UUID"
          },
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "조회할 프로젝트 UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "메시지 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "projectId",
                    "messages"
                  ],
                  "properties": {
                    "projectId": {
                      "type": "string",
                      "format": "uuid",
                      "example": "e1f2a3b4-c5d6-7890-efab-cd1234567890"
                    },
                    "messages": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Message"
                      },
                      "description": "sentAt 오름차순 정렬, 최대 200건"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "인증이 필요합니다."
                }
              }
            }
          },
          "403": {
            "description": "접근 권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "해당 팀에 접근 권한이 없습니다."
                }
              }
            }
          },
          "404": {
            "description": "프로젝트 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "프로젝트를 찾을 수 없습니다."
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ProjectMessages"
        ],
        "summary": "프로젝트 채팅 메시지 전송",
        "description": "프로젝트 전용 채팅 메시지를 전송합니다. project_id는 경로에서 자동 설정됩니다.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "소속 팀 UUID"
          },
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "메시지를 전송할 프로젝트 UUID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "content"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "NORMAL",
                      "WORK_PERFORMANCE"
                    ],
                    "default": "NORMAL",
                    "example": "NORMAL"
                  },
                  "content": {
                    "type": "string",
                    "maxLength": 2000,
                    "example": "프로젝트 진행 상황 공유합니다."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "메시지 전송 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "missing_content": {
                    "value": {
                      "error": "메시지 내용은 필수입니다."
                    }
                  },
                  "content_too_long": {
                    "value": {
                      "error": "메시지는 최대 2000자까지 입력 가능합니다."
                    }
                  },
                  "invalid_type": {
                    "value": {
                      "error": "잘못된 메시지 타입입니다."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "인증이 필요합니다."
                }
              }
            }
          },
          "403": {
            "description": "접근 권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "해당 팀에 접근 권한이 없습니다."
                }
              }
            }
          },
          "404": {
            "description": "프로젝트 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "프로젝트를 찾을 수 없습니다."
                }
              }
            }
          }
        }
      }
    },
...
    "/teams/{teamId}/board": {
      "get": {
        "tags": [
          "Board"
        ],
        "summary": "자료실 글 목록 조회",
        "description": "자료실 글 목록을 조회합니다. projectId 지정 시 프로젝트 자료실, 미지정 시 팀 일자별 자료실을 반환합니다.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "소속 팀 UUID"
          },
          {
            "name": "projectId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "프로젝트 UUID. 지정 시 해당 프로젝트 자료실 조회"
          }
        ],
        "responses": {
          "200": {
            "description": "자료실 글 목록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "projectId",
                    "posts"
                  ],
                  "properties": {
                    "projectId": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true,
                      "description": "조회에 사용된 projectId (없으면 null)",
                      "example": null
                    },
                    "posts": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/BoardPost"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "인증이 필요합니다."
                }
              }
            }
          },
          "403": {
            "description": "접근 권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "해당 팀에 접근 권한이 없습니다."
                }
              }
            }
          },
          "404": {
            "description": "프로젝트 없음 (projectId 지정 시)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "프로젝트를 찾을 수 없습니다."
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Board"
        ],
        "summary": "자료실 글 작성",
        "description": "자료실에 새 게시글을 작성합니다. 선택적으로 단일 파일을 첨부할 수 있습니다. BR-09 파일 검증 규칙 적용.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "소속 팀 UUID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "게시글 작성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardPost"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "missing_title": {
                    "value": {
                      "error": "제목은 필수입니다."
                    }
                  },
                  "title_too_long": {
                    "value": {
                      "error": "제목은 최대 200자까지 입력 가능합니다."
                    }
                  },
                  "content_too_long": {
                    "value": {
                      "error": "본문은 최대 20000자까지 입력 가능합니다."
                    }
                  },
                  "invalid_mime": {
                    "value": {
                      "error": "허용되지 않는 파일 형식입니다."
                    }
                  },
                  "file_too_large": {
                    "value": {
                      "error": "파일 크기는 최대 10MB입니다."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "인증이 필요합니다."
                }
              }
            }
          },
          "403": {
            "description": "접근 권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "해당 팀에 접근 권한이 없습니다."
                }
              }
            }
          },
          "404": {
            "description": "프로젝트 없음 (projectId 지정 시)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "프로젝트를 찾을 수 없습니다."
                }
              }
            }
          },
          "413": {
            "description": "요청 본문 크기 초과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "요청 본문이 너무 큽니다. 첨부파일은 최대 10MB."
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}